backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
cryptography>=41.0.3
SQLAlchemy>=2.0.20
psycopg2-binary>=2.9.7
secure-smtplib>=0.1.1
imapclient>=3.0.0
pyzmail36>=1.0.4
PyJWT>=2.8.0
supabase>=2.3.0
python-multipart>=0.0.9
PyMuPDF>=1.23.0
aiosmtplib>=3.0.0
pandas>=2.0.0
groq>=0.20.0
langchain-groq>=0.1.9
httpx>=0.28.1
langchain-community>=0.3.0
pinecone-client>=2.2.4
celery>=5.3.0
langchain-classic>=0.2.0
langchain-text-splitters>=0.0.1
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.0
python-docx>=0.8.11
